Make DemoDriverTestTrait.
