+++ module/VuFind/tests/integration-tests/src/VuFindTest/Mink/IlsActionsTest.php
@@ -41,6 +41,7 @@ use Behat\Mink\Element\Element;
 class IlsActionsTest extends \VuFindTest\Unit\MinkTestCase
 {
     use \VuFindTest\Unit\UserCreationTrait;
+    use \VuFindTest\Unit\DemoDriverTestTrait;
 
     /**
      * Standard setup method.
@@ -79,93 +80,6 @@ class IlsActionsTest extends \VuFindTest\Unit\MinkTestCase
                 'title_level_holds_mode' => 'driver',
                 'renewals_enabled' => true,
             ]
-        ];
-    }
-
-    /**
-     * Get transaction JSON for Demo.ini.
-     *
-     * @param string $bibId Bibliographic record ID to create fake item info for.
-     *
-     * @return array
-     */
-    protected function getFakeTransactions($bibId)
-    {
-        $rawDueDate = strtotime("now +5 days");
-        return json_encode(
-            [
-                [
-                    'duedate' => $rawDueDate,
-                    'rawduedate' => $rawDueDate,
-                    'dueStatus' => 'due',
-                    'barcode' => 1234567890,
-                    'renew'   => 0,
-                    'renewLimit' => 1,
-                    'request' => 0,
-                    'id' => $bibId,
-                    'source' => 'Solr',
-                    'item_id' => 0,
-                    'renewable' => true,
-                ]
-            ]
-        );
-    }
-
-    /**
-     * Get Demo.ini override settings for testing ILS functions.
-     *
-     * @param string $bibId Bibliographic record ID to create fake item info for.
-     *
-     * @return array
-     */
-    public function getDemoIniOverrides($bibId = 'testsample1')
-    {
-        return [
-            'Records' => [
-                'transactions' => $this->getFakeTransactions($bibId),
-            ],
-            'Failure_Probabilities' => [
-                'cancelHolds' => 0,
-                'cancelILLRequests' => 0,
-                'cancelStorageRetrievalRequests' => 0,
-                'checkILLRequestIsValid' => 0,
-                'checkRenewBlock' => 0,
-                'checkRequestIsValid' => 0,
-                'checkStorageRetrievalRequestIsValid' => 0,
-                'getAccountBlocks' => 0,
-                'getDefaultRequestGroup' => 0,
-                'getHoldDefaultRequiredDate' => 0,
-                'getRequestBlocks' => 0,
-                'placeHold' => 0,
-                'placeILLRequest' => 0,
-                'placeStorageRetrievalRequest' => 0,
-                'renewMyItems' => 0,
-            ],
-            'Holdings' => [$bibId => json_encode([$this->getFakeItem()])],
-            'Users' => ['catuser' => 'catpass'],
-        ];
-    }
-
-    /**
-     * Get a fake item record for inclusion in the Demo driver configuration.
-     *
-     * @return array
-     */
-    public function getFakeItem()
-    {
-        return [
-            'barcode'      => '12345678',
-            'availability' => true,
-            'status'       => 'Available',
-            'location'     => 'Test Location',
-            'locationhref' => false,
-            'reserve'      => 'N',
-            'callnumber'   => 'Test Call Number',
-            'duedate'      => '',
-            'is_holdable'  => true,
-            'addLink'      => true,
-            'addStorageRetrievalRequestLink' => 'check',
-            'addILLRequestLink' => 'check',
         ];
     }
